requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
chardet==5.2.0
Django==4.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.1
django-leaflet==0.30.1
djangorestframework==3.13.0
djangorestframework-gis==0.18
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.3
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
reportlab==4.2.2
sqlparse==0.4.4
typing_extensions==4.12.2
uritemplate==4.1.1